requirements.txt
python-telegram-bot==21.10
pymongo==4.10.1
rapidfuzz==3.10.1
aiohttp==3.11.11
